requirements.txt
#devops
black==22.3.0
pytest==7.1.3
pytest-cov==4.0.0
#ruff and tabulate
ruff==0.6.5
tabulate==0.9.0
# databricks req
pandas==2.2.2
python-dotenv==1.0.1
databricks-sql-connector==3.4.0
